+++ lib/presentation/home/example/home_example.dart
@@ -1,5 +1,6 @@
 import 'package:earlybuddy/domain/repository/ebauth/ebauth_repository.dart';
 import 'package:earlybuddy/presentation/home/view/home_view.dart';
+import 'package:earlybuddy/shared/eb_uikit/eb_uikit.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
@@ -12,8 +13,9 @@ class HomeExample extends StatelessWidget {
   Widget build(BuildContext context) {
     return RepositoryProvider(
       create: (context) => EBAuthRepository(),
-      child: const MaterialApp(
-        home: MockHomeView(),
+      child: MaterialApp(
+        theme: EBTheme.light(),
+        home: const MockHomeView(),
       ),
     );
   }
