+++ lib/shared/eb_uikit/sources/eb_sources.dart
@@ -3,14 +3,15 @@ import 'package:earlybuddy/shared/eb_uikit/resources/eb_resources.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
-part 'eb_textfield.dart';
-part 'eb_button.dart';
-part 'eb_wave_background.dart';
-part 'eb_roundrect_form.dart';
+part 'textfield.dart';
+part 'button.dart';
+part 'wave_background.dart';
+part 'roundrect_form.dart';
 part 'time.dart';
 part 'hexcolor.dart';
 part 'pair.dart';
 part 'dashline.dart';
 part 'scrollwithheader.dart';
 part 'expandedsection.dart';
-part 'eb_alert.dart';
+part 'alert.dart';
+part 'theme.dart';
