+++ lib/shared/eb_uikit/sources/scrollwithheader.dart
@@ -1,5 +1,3 @@
-// shrinkWrap: true,
-//                 controller: ModalScrollController.of(childContext2),
 part of 'eb_sources.dart';
 
 class ScrollWithHeader extends StatelessWidget {
